reconrisk/requirements.txt
rich>=13.0
requests>=2.28
dnspython>=2.3
